Adding ability to restore database from dump

+++ src/index.ts
@@ -2,6 +2,8 @@ import * as db from './connector'
 import * as api from './api'
 import * as sample from './sample';
 import * as cluster from 'cluster';
+import * as fs from 'fs';
+import * as cp from 'child_process';
 
 if (cluster.isMaster) {
   initMater()
@@ -13,9 +15,21 @@ async function initMater() {
   try {
     if (process.env.NODE_ENV == "development") {
       console.info("Connecting to database in DEVELOPMENT mode")
-      await db.connection.getQueryInterface().dropAllTables()
-      await db.migrate()
-      await sample.createEmptyData()
+      if (process.env.RECREATE_DB == "true") {
+        await db.connection.getQueryInterface().dropAllTables()
+        await db.connection.getQueryInterface().dropAllSchemas()
+        if (fs.existsSync("./dumps/pgdump.bin")) {
+          try {
+            cp.execSync("pg_restore --verbose --clean --no-acl --no-owner -h localhost -U steve -d postgres ./dumps/pgdump.bin", { stdio: 'inherit' })
+          } catch (e) {
+            console.warn(e);
+          }
+          await db.migrate()
+        } else {
+          await db.migrate()
+          await sample.createEmptyData()
+        }
+      }
     } else {
       console.info("Connecting to database in RELEASE mode")
       await db.migrate()
